Extract design-system alert test helper
